Added render state cache and flat hash maps

+++ BlueberryEngine/src/Blueberry/Graphics/GfxShader.h
@@ -8,9 +8,9 @@ namespace Blueberry
 		virtual ~GfxShader() = default;
 
 	protected:
-		std::unordered_map<size_t, uint32_t> m_ConstantBufferSlots;
-		std::unordered_map<size_t, std::pair<uint32_t, uint32_t>> m_StructuredBufferSlots;
-		std::unordered_map<size_t, std::pair<uint32_t, uint32_t>> m_TextureSlots;
+		std::unordered_map<size_t, uint8_t> m_ConstantBufferSlots;
+		std::unordered_map<size_t, std::pair<uint8_t, uint8_t>> m_StructuredBufferSlots;
+		std::unordered_map<size_t, std::pair<uint8_t, uint8_t>> m_TextureSlots;
 
 		friend struct GfxDrawingOperation;
 		friend class Material;
